feat: add playback speed control pane and related functionality
- Introduced a new SpeedPopup widget for adjusting playback speed.
- Added isSpeedPaneOpened observable in PlayerController to manage the visibility of the speed control pane.
- Updated BottomControls and various themes to toggle the speed pane.
- Removed the old playback speed dialog and integrated speed adjustment directly into the new pane.
- Enhanced UI for speed selection with presets and options to add custom speeds.
