+++ lib/screens/anime/watch/controls/themes/player_control_themes/netflix_desktop_player_theme.dart.dart
@@ -3,7 +3,6 @@
 import 'package:anymex/screens/anime/watch/controller/player_controller.dart';
 import 'package:anymex/screens/anime/watch/controls/themes/player_control_themes/netflix_shared.dart';
 import 'package:anymex/screens/anime/watch/controls/themes/setup/player_control_theme.dart';
-import 'package:anymex/screens/anime/watch/controls/widgets/bottom_sheet.dart';
 import 'package:anymex/screens/anime/watch/controls/widgets/progress_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -156,7 +155,8 @@ class NetflixDesktopPlayerControlTheme extends PlayerControlTheme {
                         NFDesktopButton(
                           icon: Symbols.subtitles_rounded,
                           size: 30,
-                          onTap: () => controller.isTracksPaneOpened.value = !controller.isTracksPaneOpened.value,
+                          onTap: () => controller.isTracksPaneOpened.value =
+                              !controller.isTracksPaneOpened.value,
                         ),
                         const SizedBox(width: 20),
                         NFDesktopButton(
