+++ lib/screens/anime/watch/controls/themes/player_control_themes/netflix_mobile_player_theme.dart
@@ -3,7 +3,6 @@
 import 'package:anymex/screens/anime/watch/controller/player_controller.dart';
 import 'package:anymex/screens/anime/watch/controls/themes/player_control_themes/netflix_shared.dart';
 import 'package:anymex/screens/anime/watch/controls/themes/setup/player_control_theme.dart';
-import 'package:anymex/screens/anime/watch/controls/widgets/bottom_sheet.dart';
 import 'package:anymex/screens/anime/watch/controls/widgets/progress_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -252,7 +251,9 @@ class NetflixMobilePlayerControlTheme extends PlayerControlTheme {
                               NFLabeledButton(
                                 icon: Symbols.subtitles_rounded,
                                 label: 'Audio & Subtitles',
-                                onTap: () => controller.isTracksPaneOpened.value = !controller.isTracksPaneOpened.value,
+                                onTap: () =>
+                                    controller.isTracksPaneOpened.value =
+                                        !controller.isTracksPaneOpened.value,
                               ),
                               Obx(() => NFLabeledButton(
                                     icon: Icons.skip_next_rounded,
