+++ lib/screens/anime/watch/controls/themes/player_control_themes/netflix_shared.dart
@@ -304,7 +304,7 @@ class NFMoreSheet extends StatelessWidget {
               label: 'Playback speed',
               onTap: () {
                 Get.back();
-                PlayerBottomSheets.showPlaybackSpeed(ctx, controller);
+                controller.isSpeedPaneOpened.value = true;
               },
             ),
             NFSheetTile(
